backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pymysql 
passlib[bcrypt] 
pyjwt 
